# docs/swagger.yaml
basePath: /
definitions:
  error.APIError:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a HUB API.
  title: HUB API
  version: 0.0.1
paths:
...
  /api/getAccount:
    get:
      consumes:
      - application/json
      description: get account information for a specific owner
      parameters:
      - description: account owner address
        in: query
        name: owner
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: get account information
      tags:
      - account
  /api/getBucket:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information for a specific bucket including metadata,
        storage statistics, and configuration settings. Returns bucket display information
        with owner details, creation time, and current status.
      parameters:
      - description: Account owner address (wallet address or account identifier)
        in: query
        name: owner
        type: string
      - description: Bucket name (unique identifier within the owner's account)
        in: query
        name: bucket
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bucket information including metadata, statistics, and configuration
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            $ref: '#/definitions/error.APIError'
        "404":
          description: Bucket not found
          schema:
            $ref: '#/definitions/error.APIError'
        "599":
          description: Internal server error
          schema:
            $ref: '#/definitions/error.APIError'
      summary: Get bucket information
      tags:
      - bucket
...
  /api/getNeedle:
    get:
      consumes:
      - application/json
      description: get needle information for a specific owner
      parameters:
      - default: '""'
        description: owner address
        in: query
        name: owner
        type: string
      - default: '""'
        description: bucket name
        in: query
        name: bucket
        type: string
      - default: '""'
        description: needle name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: get needle information
      tags:
      - needle
  /api/getVolume:
    get:
      consumes:
      - application/json
      description: get volume information for a specific owner
      parameters:
      - description: owner address
        in: query
        name: owner
        type: string
      - description: file name
        in: query
        name: file
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: get volume information
      tags:
      - volume
  /api/listAccount:
    get:
      consumes:
      - application/json
      description: get a list of accounts with pagination support
      parameters:
      - default: 0
        description: pagination offset
        in: query
        name: offset
        type: integer
      - default: 32
        description: number of items per page
        in: query
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: list accounts
      tags:
      - account
    post:
      consumes:
      - application/json
      description: get a list of accounts with pagination support
      parameters:
      - default: 0
        description: pagination offset
        in: formData
        name: offset
        type: integer
      - default: 32
        description: number of items per page
        in: formData
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: list accounts
      tags:
      - account
  /api/listBucket:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of buckets. If owner is specified, returns
        buckets owned by that account. If owner is empty, returns all buckets in the
        system. Supports pagination with configurable offset and page size. Returns
        bucket display information including names, creation times, and basic statistics.
      parameters:
      - default: '""'
        description: Owner address to filter buckets (leave empty to list all buckets)
        in: query
        name: owner
        type: string
      - default: 0
        description: Pagination offset - number of items to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: 32
        description: Number of items per page (max 100)
        in: query
        maximum: 100
        minimum: 1
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of buckets with metadata and statistics
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid pagination parameters
          schema:
            $ref: '#/definitions/error.APIError'
        "599":
          description: Internal server error
          schema:
            $ref: '#/definitions/error.APIError'
      summary: List buckets with pagination
      tags:
      - bucket
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Retrieve a paginated list of buckets using POST method. If owner
        is specified, returns buckets owned by that account. If owner is empty, returns
        all buckets in the system. Supports pagination with configurable offset and
        page size. Returns bucket display information including names, creation times,
        and basic statistics. This POST variant is useful for complex filtering scenarios
        or when URL length limits are a concern.
      parameters:
      - default: '""'
        description: Owner address to filter buckets (leave empty to list all buckets)
        in: formData
        name: owner
        type: string
      - default: 0
        description: Pagination offset - number of items to skip
        in: formData
        minimum: 0
        name: offset
        type: integer
      - default: 32
        description: Number of items per page (max 100)
        in: formData
        maximum: 100
        minimum: 1
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of buckets with metadata and statistics
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid pagination parameters
          schema:
            $ref: '#/definitions/error.APIError'
        "599":
          description: Internal server error
          schema:
            $ref: '#/definitions/error.APIError'
      summary: List buckets with pagination (POST)
      tags:
      - bucket
...
  /api/listNeedle:
    get:
      consumes:
      - application/json
      description: get a list of needles for a specific owner with pagination support
      parameters:
      - default: '""'
        description: owner address
        in: query
        name: owner
        type: string
      - default: '""'
        description: bucket name
        in: query
        name: bucket
        type: string
      - default: 0
        description: pagination offset
        in: query
        name: offset
        type: integer
      - default: 32
        description: number of items per page
        in: query
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: list needles
      tags:
      - needle
    post:
      consumes:
      - application/json
      description: get a list of needles for a specific owner with pagination support
      parameters:
      - default: '""'
        description: owner address
        in: formData
        name: owner
        type: string
      - default: '""'
        description: bucket name
        in: formData
        name: bucket
        type: string
      - default: 0
        description: pagination offset
        in: formData
        name: offset
        type: integer
      - default: 32
        description: number of items per page
        in: formData
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: list needles
      tags:
      - needle
  /api/listVolume:
    get:
      consumes:
      - application/json
      description: get a list of volumes for a specific owner with pagination support
      parameters:
      - description: owner address
        in: query
        name: owner
        type: string
      - default: 0
        description: pagination offset
        in: query
        name: offset
        type: integer
      - default: 32
        description: number of items per page
        in: query
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: list volumes
      tags:
      - volume
    post:
      consumes:
      - application/json
      description: get a list of volumes for a specific owner with pagination support
      parameters:
      - description: owner address
        in: formData
        name: owner
        type: string
      - default: 0
        description: pagination offset
        in: formData
        name: offset
        type: integer
      - default: 32
        description: number of items per page
        in: formData
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "599":
          description: ""
          schema:
            $ref: '#/definitions/error.APIError'
      summary: list volumes
      tags:
      - volume
...
swagger: "2.0"